move filter pr or issues logic to utlils helper
