+++ controllers/users.js
@@ -8,11 +8,10 @@ const { logType } = require("../constants/logs");
 const { fetch } = require("../utils/fetch");
 const logger = require("../utils/logger");
 const obfuscate = require("../utils/obfuscate");
-const githubService = require("../services/githubService");
 const { getPaginationLink, getFilteredUsers } = require("../utils/users");
-const { getQualifiers, getDateTimeRangeForPRs } = require("../utils/helper");
-const { fetchMultiplePageResults } = require("../utils/fetchMultiplePageResults");
+const { getQualifiers } = require("../utils/helper");
 const { SOMETHING_WENT_WRONG, INTERNAL_SERVER_ERROR } = require("../constants/errorMessages");
+const { getFilteredPRsOrIssues } = require("../utils/github");
 
 const verifyUser = async (req, res) => {
   const userId = req.userData.id;
@@ -68,63 +67,11 @@ const getUserById = async (req, res) => {
 
 const getUsers = async (req, res) => {
   try {
-    let allPRs = [];
     const query = req.query.q;
     const qualifiers = getQualifiers(query);
 
     if (qualifiers?.filterBy) {
-      const { sortBy = "RECENT_FIRST", filterBy } = qualifiers;
-      const order = sortBy === "RECENT_FIRST" ? "desc" : "asc";
-
-      const startDate = qualifiers?.startDate;
-      const endDate = qualifiers?.endDate;
-      const dateTime = getDateTimeRangeForPRs(startDate, endDate);
-
-      const searchParams = {}; // searchParams used to create list of params to create github API URL
-      const resultOptions = { order }; // resultOptions is used for ordering, searching within date-time range and pagination of results
-
-      if (filterBy === "OPEN_PRS") {
-        if (dateTime) {
-          searchParams.created = dateTime;
-        }
-
-        allPRs = await fetchMultiplePageResults(githubService.fetchOpenPRs, {
-          searchParams,
-          resultOptions,
-        });
-      }
-      if (filterBy === "MERGED_PRS") {
-        if (dateTime) {
-          searchParams.merged = dateTime;
-        }
-
-        allPRs = await fetchMultiplePageResults(githubService.fetchMergedPRs, {
-          searchParams,
-          resultOptions,
-        });
-      }
-
-      if (filterBy === "OPEN_ISSUES") {
-        if (dateTime) {
-          searchParams.created = dateTime;
-        }
-
-        allPRs = await fetchMultiplePageResults(githubService.fetchOpenIssues, {
-          searchParams,
-          resultOptions,
-        });
-      }
-
-      if (filterBy === "CLOSED_ISSUES") {
-        if (dateTime) {
-          searchParams.closed = dateTime;
-        }
-
-        allPRs = await fetchMultiplePageResults(githubService.fetchClosedIssues, {
-          searchParams,
-          resultOptions,
-        });
-      }
+      const allPRs = await getFilteredPRsOrIssues(qualifiers);
 
       const { allUsers } = await userQuery.fetchAllUsers();
